# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TAPER",
        "version": "0.1"
    },
    "paths": {
        "/api/admin-panel/get-banners": {
            "get": {
                "tags": [
                    "Админ панель"
                ],
                "summary": "Получение списка баннеров",
                "operationId": "5b62c906ddf7133dcb212332db4390cc",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/admin-panel/banner/{id}": {
            "get": {
                "tags": [
                    "Админ панель"
                ],
                "summary": "Получение баннера по id",
                "operationId": "c16d0c70888afc9be6c800d617c881a5",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/admin-panel/create-banner": {
            "post": {
                "tags": [
                    "Админ панель"
                ],
                "summary": "Создание банера",
                "operationId": "c0486273d1336a4c1b974c897321cac3",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "file",
                                    "dateStart",
                                    "dateStop"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Имя",
                                        "type": "string"
                                    },
                                    "file": {
                                        "description": "file",
                                        "type": "file"
                                    },
                                    "dateStart": {
                                        "description": "file",
                                        "type": "string"
                                    },
                                    "dateStop": {
                                        "description": "file",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin-panel/update-banner/{id}": {
            "post": {
                "tags": [
                    "Админ панель"
                ],
                "summary": "Изменение банера",
                "operationId": "f8322557ac2cf2550aa20e70009f407c",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "file",
                                    "dateStart",
                                    "dateStop"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Имя",
                                        "type": "string"
                                    },
                                    "file": {
                                        "description": "file",
                                        "type": "file"
                                    },
                                    "dateStart": {
                                        "description": "file",
                                        "type": "string"
                                    },
                                    "dateStop": {
                                        "description": "file",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin-panel/get-film-copies": {
            "get": {
                "tags": [
                    "Админ панель"
                ],
                "summary": "Получение списка баннеров",
                "operationId": "ad7f92e8053bf38bcb39dcff6c09797a",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/admin-panel/get-film-copy/{id}": {
            "get": {
                "tags": [
                    "Админ панель"
                ],
                "summary": "Получение баннера по id",
                "operationId": "c46f3194d2c7e33f8de98a1a92ff3f96",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/admin-panel/update-film-copy/{id}": {
            "post": {
                "tags": [
                    "Админ панель"
                ],
                "summary": "Store",
                "operationId": "63a1f6e5c3fa9e399e623fdbc1c96dd7",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "file",
                                    "dateStart",
                                    "dateStop"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Имя",
                                        "type": "string"
                                    },
                                    "file": {
                                        "description": "file",
                                        "type": "file"
                                    },
                                    "dateStart": {
                                        "description": "file",
                                        "type": "string"
                                    },
                                    "dateStop": {
                                        "description": "file",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/user/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Auth manager",
                "operationId": "0defdf852b4dad70e33dde9ee4cf6380",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password"
                                ],
                                "properties": {
                                    "login": {
                                        "description": "логин пользователя",
                                        "type": "string",
                                        "example": "123123"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}